correct a bug with linux 64bit system. A 'C' thread function needed a long and python arg passing default to int. Forced to long wiht ctypes.c_long()

+++ IPython/gui/wx/thread_ex.py
@@ -12,7 +12,7 @@ def _async_raise(tid, exctype):
     """raises the exception, performs cleanup if needed"""
     if not inspect.isclass(exctype):
         raise TypeError("Only types can be raised (not instances)")
-    res = ctypes.pythonapi.PyThreadState_SetAsyncExc(tid, ctypes.py_object(exctype))
+    res = ctypes.pythonapi.PyThreadState_SetAsyncExc(ctypes.c_long(tid), ctypes.py_object(exctype))
     if res == 0:
         raise ValueError("invalid thread id")
     elif res != 1:
